refactor(core): split seven oversize modules under the 500-LOC ratchet
Pure mechanical splits, no behavior change; original module paths keep
their exact export surfaces so no importer changes:

- theme-builtins.ts (775->32) -> cohort tables in theme-builtins-{core,
  ggthemes,publication}.ts + legacy overlay builder; insertion order now
  pinned by an exact Object.keys assertion
- stats/loess.ts (663->497) -> loess-{linear-system,interpolate,statistics}.ts
- svg-live/patch-marks.ts (630->87) -> patch-{attrs,points,paths,rects,
  segments}.ts; BatchPatchContext lives in patch-attrs (no cycles)
- layout/band-guide.ts (527->309) -> band-guide-rotated.ts factory owning
  the rotated/truncated/thinned leg; escalation ladder untouched
- candidate-store-indexes.ts (720->455) -> index types + coincidence +
  lazy axis-group modules; only final arrays cross the retained-memory
  boundary (bench:memory:check green)
- table.ts (639->25) -> table-parse-lite.ts + table-column.ts; table.ts
  is now an explicit barrel, ./table.js importers unchanged
- pipeline/build-candidates.ts (703->165) -> candidate-source-{state,
  values,resolver,columns}.ts; callback/columnar parity preserved verbatim
- scales-m1 theme test split into 4 scenario files (518->80 + 3 new)

Ratchet entries pruned for all seven (71 -> 64). Verification: bun test
packages/core (2484 pass), packages/svelte svg-live-patch (30 pass across
chromium/firefox/webkit), bun run check, knip, lint:type-aware,
bench:memory:check, max-loc --all.
