+++ packages/core/src/theme-builtins-core.ts
@@ -1,0 +1,268 @@
+/**
+ * Core built-in themes: default/hrbr lineage plus the ggplot2-derived ports
+ * (ggplot2/grey/gray, classic, bw, fivethirtyeight, economist, tufte,
+ * linedraw, void). Assembled in order by theme-builtins.ts.
+ */
+import type { ThemeName } from "@ggsvelte/spec";
+import { themed, type ThemeTokens } from "./theme-construct.js";
+
+/**
+ * Grey-panel ggplot2 complete-theme tokens. Shared by registered names
+ * `ggplot2`, `grey`, and `gray` (UK/US theme_grey / theme_gray aliases, #824).
+ * LEGACY_BUILTIN_THEMES spreads BUILTIN_THEMES so the aliases inherit there too.
+ */
+const GGPLOT2_GREY = themed({
+  ink: "#333333",
+  panel: "#ebebeb",
+  grid: "#ffffff",
+  axisText: "#4d4d4d",
+  tickColor: "#333333",
+  fontSize: 11,
+  axisTextSize: 12,
+  fontWeight: 400,
+  titleSize: 13.2,
+  subtitleSize: 11,
+  axisTitleSize: 11,
+  captionSize: 8.8,
+  stripSize: 8.8,
+  tickWidth: 0.5,
+  tickLength: 3.67,
+  gridWidth: 0.5,
+  ticksX: true,
+  ticksY: true,
+});
+
+/**
+ * Built-in themes for edition 2. The default deliberately follows
+ * hrbrthemes' quiet hierarchy: real typography, hairline grids, and no heavy
+ * axis frame. Named presets retain the structural contracts of their R
+ * counterparts rather than acting as color aliases.
+ */
+export const CORE_BUILTIN_THEMES = Object.freeze({
+  default: themed({}),
+  hrbr: themed({}),
+  minimal: themed({
+    ink: "#333333",
+    accent: "#4385be",
+    grid: "#ebebeb",
+    fontSize: 11,
+    // Was 8.8 — unreadable next to titles/tooltips at chart size (#753).
+    axisTextSize: 12,
+    titleSize: 15,
+    subtitleSize: 12,
+    gridWidth: 0.5,
+  }),
+  light: themed({
+    ink: "#333333",
+    grid: "#dedede",
+    axisText: "#4d4d4d",
+    axisLine: "#b3b3b3",
+    tickColor: "#b3b3b3",
+    fontSize: 11,
+    // Was 8.8 — homepage hero and light charts forced squinting (#753).
+    // Smoke VR baselines refreshed with this token change (PR #755).
+    axisTextSize: 12,
+    titleSize: 15,
+    subtitleSize: 12,
+    axisLineWidth: 0.5,
+    tickWidth: 0.25,
+    tickLength: 3.67,
+    gridWidth: 0.25,
+    ticksX: true,
+    ticksY: true,
+    showPanelBorder: true,
+  }),
+  dark: themed({
+    ink: "#e6e8eb",
+    paper: "#16181d",
+    panel: "#16181d",
+    accent: "#7ea1f0",
+    grid: "#3b3f46",
+    axisText: "#c6c9ce",
+    axisLine: "#6b717b",
+    tickColor: "#6b717b",
+    // Elevated card fill only; ink/border stay derived from foundation.
+    tooltipPaper: "#22262d",
+  }),
+  ggplot2: GGPLOT2_GREY,
+  // theme_grey / theme_gray name parity — same token map as ggplot2 (#824).
+  grey: GGPLOT2_GREY,
+  gray: GGPLOT2_GREY,
+  classic: themed({
+    ink: "#000000",
+    grid: "none",
+    axisText: "#000000",
+    axisLine: "#000000",
+    tickColor: "#000000",
+    fontSize: 11,
+    axisTextSize: 12,
+    fontWeight: 400,
+    titleSize: 15,
+    subtitleSize: 12,
+    axisTitleSize: 11,
+    axisLineWidth: 0.5,
+    tickWidth: 0.5,
+    tickLength: 3.67,
+    gridWidth: 0,
+    axisLineX: true,
+    axisLineY: true,
+    ticksX: true,
+    ticksY: true,
+    gridX: false,
+    gridY: false,
+  }),
+  // theme_bw complete-theme role (#820): white panel, grey grid, rectangular
+  // border for print/B&W reproduction. Distinct from ggplot2 (grey panel,
+  // white grid) and light (thinner chrome, lighter border).
+  bw: themed({
+    ink: "#333333",
+    paper: "#ffffff",
+    panel: "#ffffff",
+    grid: "#e5e5e5",
+    axisText: "#4d4d4d",
+    axisLine: "#333333",
+    tickColor: "#333333",
+    panelBorder: "#333333",
+    fontSize: 11,
+    axisTextSize: 12,
+    fontWeight: 400,
+    titleSize: 13.2,
+    subtitleSize: 11,
+    axisTitleSize: 11,
+    captionSize: 8.8,
+    stripSize: 8.8,
+    axisLineWidth: 0.5,
+    tickWidth: 0.5,
+    tickLength: 3.67,
+    gridWidth: 0.5,
+    panelBorderWidth: 0.5,
+    ticksX: true,
+    ticksY: true,
+    axisLineX: false,
+    axisLineY: false,
+    showPanelBorder: true,
+  }),
+  few: themed({
+    ink: "#333333",
+    accent: "#5da5da",
+    grid: "none",
+    axisText: "#4d4d4d",
+    fontSize: 12,
+    axisTextSize: 12,
+    fontWeight: 400,
+    titleSize: 14.4,
+    subtitleSize: 12,
+    axisTitleSize: 12,
+    axisLine: "#4d4d4d",
+    tickColor: "#4d4d4d",
+    panelBorder: "#4d4d4d",
+    axisLineWidth: 0.5,
+    tickWidth: 0.5,
+    tickLength: 3,
+    gridWidth: 0,
+    ticksX: true,
+    ticksY: true,
+    gridX: false,
+    gridY: false,
+    showPanelBorder: true,
+  }),
+  clean: themed({
+    accent: "#5da5da",
+    grid: "#b3b3b3",
+    axisLine: "#4d4d4d",
+    tickColor: "#4d4d4d",
+    axisLineWidth: 0.5,
+    tickWidth: 0.5,
+    tickLength: 3,
+    gridWidth: 0.4,
+    gridDasharray: "1.5 2.5",
+    axisLineX: true,
+    axisLineY: true,
+    ticksX: true,
+    ticksY: true,
+    gridX: false,
+  }),
+  fivethirtyeight: themed({
+    ink: "#3c3c3c",
+    paper: "#f0f0f0",
+    panel: "#f0f0f0",
+    accent: "#008fd5",
+    grid: "#ffffff",
+    axisText: "#3c3c3c",
+    gridWidth: 0.5,
+    // White card over grey paper (theme highlight tone); soft hairline (fill is primary).
+    tooltipPaper: "#ffffff",
+    tooltipBorder: "#d0d0d0",
+  }),
+  economist: themed({
+    ink: "#014d64",
+    paper: "#d5e4eb",
+    panel: "#d5e4eb",
+    accent: "#ed111a",
+    grid: "#ffffff",
+    axisText: "#014d64",
+    tickColor: "#6794a7",
+    tickWidth: 0.5,
+    tickLength: 4,
+    gridWidth: 0.5,
+    ticksX: true,
+    // Same-hue lighter tip + tick-color border (not pure white wash).
+    tooltipPaper: "#eef5f8",
+    tooltipBorder: "#6794a7",
+  }),
+  tufte: themed({
+    ink: "#111111",
+    accent: "#111111",
+    grid: "none",
+    gridWidth: 0,
+    gridX: false,
+    gridY: false,
+  }),
+  // theme_linedraw: white panel + black grid/border/ticks (line-art / B&W print).
+  // Hairline black grid (0.3) avoids solid graph paper; panel border carries the
+  // frame (axisLineX/Y stay false from HRBR_BASE). Accent stays monochrome so
+  // unmapped marks and focus chrome do not introduce a brand hue.
+  linedraw: themed({
+    ink: "#000000",
+    paper: "#ffffff",
+    panel: "#ffffff",
+    accent: "#000000",
+    grid: "#000000",
+    axisText: "#000000",
+    axisLine: "#000000",
+    tickColor: "#000000",
+    panelBorder: "#000000",
+    fontSize: 11,
+    axisTextSize: 12,
+    fontWeight: 400,
+    titleSize: 13.2,
+    subtitleSize: 11,
+    axisTitleSize: 11,
+    captionSize: 8.8,
+    stripSize: 8.8,
+    tickWidth: 0.5,
+    tickLength: 3.67,
+    gridWidth: 0.3,
+    panelBorderWidth: 0.5,
+    ticksX: true,
+    ticksY: true,
+    showPanelBorder: true,
+  }),
+  // ggplot2 theme_void: no axes, grid, or panel chrome; marks (and legends) remain.
+  void: themed({
+    paper: "none",
+    panel: "none",
+    grid: "none",
+    gridWidth: 0,
+    gridX: false,
+    gridY: false,
+    axisLineX: false,
+    axisLineY: false,
+    ticksX: false,
+    ticksY: false,
+    labelsX: false,
+    labelsY: false,
+    showPanelBorder: false,
+    tickLength: 0,
+  }),
+} satisfies Partial<Record<ThemeName, ThemeTokens>>);
