+++ packages/core/src/theme-builtins.ts
@@ -1,775 +1,31 @@
 /**
  * Built-in theme tables and token construction.
  *
- * Holds the frozen BUILTIN_THEMES / LEGACY_BUILTIN_THEMES registries.
- * Token construction lives in theme-construct.ts; resolution in theme-resolve.ts.
+ * Holds the frozen BUILTIN_THEMES / LEGACY_BUILTIN_THEMES registries, assembled
+ * from the cohort tables in theme-builtins-{core,ggthemes,publication}.ts; the
+ * edition-1 overlay lives in theme-builtins-legacy.ts. Token construction is
+ * theme-construct.ts; resolution is theme-resolve.ts.
  */
 import type { ThemeName } from "@ggsvelte/spec";
+import type { ThemeTokens } from "./theme-construct.js";
+import { CORE_BUILTIN_THEMES } from "./theme-builtins-core.js";
+import { SOLARIZED_BUILTIN_THEMES } from "./theme-builtins-ggthemes.js";
+import { PUBLICATION_BUILTIN_THEMES } from "./theme-builtins-publication.js";
+import { legacyBuiltinThemes } from "./theme-builtins-legacy.js";
 
 export type { ThemeTokens } from "./theme-construct.js";
 export { themed } from "./theme-construct.js";
-import { LEGACY_BASE_FOUNDATION, themed, type ThemeTokens } from "./theme-construct.js";
 
 /**
- * Grey-panel ggplot2 complete-theme tokens. Shared by registered names
- * `ggplot2`, `grey`, and `gray` (UK/US theme_grey / theme_gray aliases, #824).
- * LEGACY_BUILTIN_THEMES spreads BUILTIN_THEMES so the aliases inherit there too.
- */
-const GGPLOT2_GREY = themed({
-  ink: "#333333",
-  panel: "#ebebeb",
-  grid: "#ffffff",
-  axisText: "#4d4d4d",
-  tickColor: "#333333",
-  fontSize: 11,
-  axisTextSize: 12,
-  fontWeight: 400,
-  titleSize: 13.2,
-  subtitleSize: 11,
-  axisTitleSize: 11,
-  captionSize: 8.8,
-  stripSize: 8.8,
-  tickWidth: 0.5,
-  tickLength: 3.67,
-  gridWidth: 0.5,
-  ticksX: true,
-  ticksY: true,
-});
-
-/**
- * Built-in themes for edition 2. The default deliberately follows
- * hrbrthemes' quiet hierarchy: real typography, hairline grids, and no heavy
- * axis frame. Named presets retain the structural contracts of their R
- * counterparts rather than acting as color aliases.
+ * Built-in themes for edition 2, frozen in the documented registration order:
+ * core defaults, ggthemes ports (Stata/Solarized/Economist), then publication
+ * styles through the pinned test chrome.
  */
 export const BUILTIN_THEMES: Readonly<Record<ThemeName, ThemeTokens>> = Object.freeze({
-  default: themed({}),
-  hrbr: themed({}),
-  minimal: themed({
-    ink: "#333333",
-    accent: "#4385be",
-    grid: "#ebebeb",
-    fontSize: 11,
-    // Was 8.8 — unreadable next to titles/tooltips at chart size (#753).
-    axisTextSize: 12,
-    titleSize: 15,
-    subtitleSize: 12,
-    gridWidth: 0.5,
-  }),
-  light: themed({
-    ink: "#333333",
-    grid: "#dedede",
-    axisText: "#4d4d4d",
-    axisLine: "#b3b3b3",
-    tickColor: "#b3b3b3",
-    fontSize: 11,
-    // Was 8.8 — homepage hero and light charts forced squinting (#753).
-    // Smoke VR baselines refreshed with this token change (PR #755).
-    axisTextSize: 12,
-    titleSize: 15,
-    subtitleSize: 12,
-    axisLineWidth: 0.5,
-    tickWidth: 0.25,
-    tickLength: 3.67,
-    gridWidth: 0.25,
-    ticksX: true,
-    ticksY: true,
-    showPanelBorder: true,
-  }),
-  dark: themed({
-    ink: "#e6e8eb",
-    paper: "#16181d",
-    panel: "#16181d",
-    accent: "#7ea1f0",
-    grid: "#3b3f46",
-    axisText: "#c6c9ce",
-    axisLine: "#6b717b",
-    tickColor: "#6b717b",
-    // Elevated card fill only; ink/border stay derived from foundation.
-    tooltipPaper: "#22262d",
-  }),
-  ggplot2: GGPLOT2_GREY,
-  // theme_grey / theme_gray name parity — same token map as ggplot2 (#824).
-  grey: GGPLOT2_GREY,
-  gray: GGPLOT2_GREY,
-  classic: themed({
-    ink: "#000000",
-    grid: "none",
-    axisText: "#000000",
-    axisLine: "#000000",
-    tickColor: "#000000",
-    fontSize: 11,
-    axisTextSize: 12,
-    fontWeight: 400,
-    titleSize: 15,
-    subtitleSize: 12,
-    axisTitleSize: 11,
-    axisLineWidth: 0.5,
-    tickWidth: 0.5,
-    tickLength: 3.67,
-    gridWidth: 0,
-    axisLineX: true,
-    axisLineY: true,
-    ticksX: true,
-    ticksY: true,
-    gridX: false,
-    gridY: false,
-  }),
-  // theme_bw complete-theme role (#820): white panel, grey grid, rectangular
-  // border for print/B&W reproduction. Distinct from ggplot2 (grey panel,
-  // white grid) and light (thinner chrome, lighter border).
-  bw: themed({
-    ink: "#333333",
-    paper: "#ffffff",
-    panel: "#ffffff",
-    grid: "#e5e5e5",
-    axisText: "#4d4d4d",
-    axisLine: "#333333",
-    tickColor: "#333333",
-    panelBorder: "#333333",
-    fontSize: 11,
-    axisTextSize: 12,
-    fontWeight: 400,
-    titleSize: 13.2,
-    subtitleSize: 11,
-    axisTitleSize: 11,
-    captionSize: 8.8,
-    stripSize: 8.8,
-    axisLineWidth: 0.5,
-    tickWidth: 0.5,
-    tickLength: 3.67,
-    gridWidth: 0.5,
-    panelBorderWidth: 0.5,
-    ticksX: true,
-    ticksY: true,
-    axisLineX: false,
-    axisLineY: false,
-    showPanelBorder: true,
-  }),
-  few: themed({
-    ink: "#333333",
-    accent: "#5da5da",
-    grid: "none",
-    axisText: "#4d4d4d",
-    fontSize: 12,
-    axisTextSize: 12,
-    fontWeight: 400,
-    titleSize: 14.4,
-    subtitleSize: 12,
-    axisTitleSize: 12,
-    axisLine: "#4d4d4d",
-    tickColor: "#4d4d4d",
-    panelBorder: "#4d4d4d",
-    axisLineWidth: 0.5,
-    tickWidth: 0.5,
-    tickLength: 3,
-    gridWidth: 0,
-    ticksX: true,
-    ticksY: true,
-    gridX: false,
-    gridY: false,
-    showPanelBorder: true,
-  }),
-  clean: themed({
-    accent: "#5da5da",
-    grid: "#b3b3b3",
-    axisLine: "#4d4d4d",
-    tickColor: "#4d4d4d",
-    axisLineWidth: 0.5,
-    tickWidth: 0.5,
-    tickLength: 3,
-    gridWidth: 0.4,
-    gridDasharray: "1.5 2.5",
-    axisLineX: true,
-    axisLineY: true,
-    ticksX: true,
-    ticksY: true,
-    gridX: false,
-  }),
-  fivethirtyeight: themed({
-    ink: "#3c3c3c",
-    paper: "#f0f0f0",
-    panel: "#f0f0f0",
-    accent: "#008fd5",
-    grid: "#ffffff",
-    axisText: "#3c3c3c",
-    gridWidth: 0.5,
-    // White card over grey paper (theme highlight tone); soft hairline (fill is primary).
-    tooltipPaper: "#ffffff",
-    tooltipBorder: "#d0d0d0",
-  }),
-  economist: themed({
-    ink: "#014d64",
-    paper: "#d5e4eb",
-    panel: "#d5e4eb",
-    accent: "#ed111a",
-    grid: "#ffffff",
-    axisText: "#014d64",
-    tickColor: "#6794a7",
-    tickWidth: 0.5,
-    tickLength: 4,
-    gridWidth: 0.5,
-    ticksX: true,
-    // Same-hue lighter tip + tick-color border (not pure white wash).
-    tooltipPaper: "#eef5f8",
-    tooltipBorder: "#6794a7",
-  }),
-  tufte: themed({
-    ink: "#111111",
-    accent: "#111111",
-    grid: "none",
-    gridWidth: 0,
-    gridX: false,
-    gridY: false,
-  }),
-  // theme_linedraw: white panel + black grid/border/ticks (line-art / B&W print).
-  // Hairline black grid (0.3) avoids solid graph paper; panel border carries the
-  // frame (axisLineX/Y stay false from HRBR_BASE). Accent stays monochrome so
-  // unmapped marks and focus chrome do not introduce a brand hue.
-  linedraw: themed({
-    ink: "#000000",
-    paper: "#ffffff",
-    panel: "#ffffff",
-    accent: "#000000",
-    grid: "#000000",
-    axisText: "#000000",
-    axisLine: "#000000",
-    tickColor: "#000000",
-    panelBorder: "#000000",
-    fontSize: 11,
-    axisTextSize: 12,
-    fontWeight: 400,
-    titleSize: 13.2,
-    subtitleSize: 11,
-    axisTitleSize: 11,
-    captionSize: 8.8,
-    stripSize: 8.8,
-    tickWidth: 0.5,
-    tickLength: 3.67,
-    gridWidth: 0.3,
-    panelBorderWidth: 0.5,
-    ticksX: true,
-    ticksY: true,
-    showPanelBorder: true,
-  }),
-  // ggplot2 theme_void: no axes, grid, or panel chrome; marks (and legends) remain.
-  void: themed({
-    paper: "none",
-    panel: "none",
-    grid: "none",
-    gridWidth: 0,
-    gridX: false,
-    gridY: false,
-    axisLineX: false,
-    axisLineY: false,
-    ticksX: false,
-    ticksY: false,
-    labelsX: false,
-    labelsY: false,
-    showPanelBorder: false,
-    tickLength: 0,
-  }),
-  // ggthemes theme_stata(scheme = "s2color"): the default Stata look —
-  // ltbluishgray (#eaf2f3) plot region around a white panel, matching
-  // ltbluishgray y-major grid (major.x and minor blank), black axis lines
-  // and ticks, no panel border. Sizes from stata_gsize ratios (base 11,
-  // axis 10, title 14, axis title 10). R's dknavy title colour folds into
-  // the single ink role (documented), and the bottom legend position is not
-  // expressible. Accent is the s2color palette's navy.
-  stata: themed({
-    ink: "#000000",
-    paper: "#eaf2f3",
-    panel: "#ffffff",
-    accent: "#1a476f",
-    grid: "#eaf2f3",
-    axisText: "#000000",
-    axisLine: "#000000",
-    tickColor: "#000000",
-    fontSize: 11,
-    axisTextSize: 10,
-    fontWeight: 400,
-    titleSize: 14,
-    subtitleSize: 11,
-    axisTitleSize: 10,
-    captionSize: 8,
-    stripSize: 12,
-    axisLineWidth: 0.5,
-    tickWidth: 0.5,
-    tickLength: 4,
-    gridWidth: 0.5,
-    axisLineX: true,
-    axisLineY: true,
-    ticksX: true,
-    ticksY: true,
-    gridX: false,
-  }),
-  // ggthemes theme_stata(scheme = "s1color"): the older Stata s1 look —
-  // white plot and panel, gs14 (#e0e0e0) y-grid, and Stata's black panel
-  // border. Accent is the s1color palette's dkgreen.
-  stata_s1color: themed({
-    ink: "#000000",
-    paper: "#ffffff",
-    panel: "#ffffff",
-    accent: "#006000",
-    grid: "#e0e0e0",
-    axisText: "#000000",
-    axisLine: "#000000",
-    tickColor: "#000000",
-    panelBorder: "#000000",
-    fontSize: 11,
-    axisTextSize: 10,
-    fontWeight: 400,
-    titleSize: 14,
-    subtitleSize: 11,
-    axisTitleSize: 10,
-    captionSize: 8,
-    stripSize: 12,
-    axisLineWidth: 0.5,
-    tickWidth: 0.5,
-    tickLength: 4,
-    gridWidth: 0.5,
-    panelBorderWidth: 0.5,
-    axisLineX: true,
-    axisLineY: true,
-    ticksX: true,
-    ticksY: true,
-    gridX: false,
-    showPanelBorder: true,
-  }),
-  // ggthemes theme_solarized (light = TRUE): theme_bw lineage on Schoonover's
-  // light rebase — base3 panel (#fdf6e3), base2 grid, base1 chrome/text,
-  // transparent plot background, blue accent. ggthemes draws the frame via
-  // panel.background colour (panel.border is blank); here showPanelBorder
-  // carries it. Single-ink model: R's darker title step (rebase0) flattens
-  // into ink (rebase01), as with the other ports.
-  solarized: themed({
-    ink: "#93a1a1",
-    paper: "none",
-    panel: "#fdf6e3",
-    accent: "#268bd2",
-    grid: "#eee8d5",
-    axisText: "#93a1a1",
-    axisLine: "#93a1a1",
-    tickColor: "#93a1a1",
-    panelBorder: "#93a1a1",
-    fontSize: 12,
-    axisTextSize: 12,
-    fontWeight: 400,
-    titleSize: 14.4,
-    subtitleSize: 12,
-    axisTitleSize: 12,
-    axisLineWidth: 0.5,
-    tickWidth: 0.5,
-    tickLength: 3.67,
-    gridWidth: 0.5,
-    panelBorderWidth: 0.5,
-    // theme_bw lineage: frame is the panel border only (axisLineX/Y stay false).
-    axisLineX: false,
-    axisLineY: false,
-    ticksX: true,
-    ticksY: true,
-    showPanelBorder: true,
-    // Tip package: base2 card; tip ink intentionally stronger than axis ink
-    // (base01 vs base1) for short-lived reading chrome.
-    tooltipPaper: "#eee8d5",
-    tooltipInk: "#586e75",
-    tooltipBorder: "#93a1a1",
-  }),
-  // ggthemes theme_solarized (light = FALSE): same geometry on the dark
-  // rebase — base03 panel (#002b36), base02 grid, base01 chrome. Accents are
-  // shared across light/dark by Solarized's design (only base tones flip).
-  solarizeddark: themed({
-    ink: "#586e75",
-    paper: "none",
-    panel: "#002b36",
-    accent: "#268bd2",
-    grid: "#073642",
-    axisText: "#586e75",
-    axisLine: "#586e75",
-    tickColor: "#586e75",
-    panelBorder: "#586e75",
-    fontSize: 12,
-    axisTextSize: 12,
-    fontWeight: 400,
-    titleSize: 14.4,
-    subtitleSize: 12,
-    axisTitleSize: 12,
-    axisLineWidth: 0.5,
-    tickWidth: 0.5,
-    tickLength: 3.67,
-    gridWidth: 0.5,
-    panelBorderWidth: 0.5,
-    // theme_bw lineage: frame is the panel border only (axisLineX/Y stay false).
-    axisLineX: false,
-    axisLineY: false,
-    ticksX: true,
-    ticksY: true,
-    showPanelBorder: true,
-    // Tip package: base02 card; tip ink brightens from axis base01 to base1.
-    tooltipPaper: "#073642",
-    tooltipInk: "#93a1a1",
-    tooltipBorder: "#586e75",
-  }),
-  // ggthemes theme_economist_white (gray_bg = TRUE): the Graphic Detail blog
-  // variant of theme_economist — white panel on a light-gray (#ebebeb) plot
-  // background with a dark-gray (#c9c9c9) major grid. R builds it as
-  // theme_economist() + surface overrides, so this port keeps the economist
-  // chrome (ink, ticks, sizes) and swaps exactly the three surfaces.
-  economist_white: themed({
-    ink: "#014d64",
-    paper: "#ebebeb",
-    panel: "#ffffff",
-    accent: "#ed111a",
-    grid: "#c9c9c9",
-    axisText: "#014d64",
-    tickColor: "#6794a7",
-    tickWidth: 0.5,
-    tickLength: 4,
-    gridWidth: 0.5,
-    ticksX: true,
-  }),
-  // ggthemes theme_solarized_2 (light = TRUE): the theme_grey-flavored
-  // Solarized variant — base2 panel (#eee8d5) with base3 grid (#fdf6e3),
-  // base1 chrome/text, transparent plot background. In ggthemes the axis
-  // line colour reads a misspelled rebase key ("reabase01") and resolves to
-  // NA, so no axis line is drawn; panel.border is blank, so no frame either.
-  // Single-ink model: R's darker title step (rebase0) flattens into ink
-  // (rebase01), as with the other ports.
-  solarized_2: themed({
-    ink: "#93a1a1",
-    paper: "none",
-    panel: "#eee8d5",
-    accent: "#268bd2",
-    grid: "#fdf6e3",
-    axisText: "#93a1a1",
-    tickColor: "#93a1a1",
-    fontSize: 12,
-    axisTextSize: 12,
-    fontWeight: 400,
-    titleSize: 14.4,
-    subtitleSize: 12,
-    axisTitleSize: 12,
-    tickWidth: 0.5,
-    tickLength: 3.67,
-    gridWidth: 0.5,
-    ticksX: true,
-    ticksY: true,
-    // Tip package: base3 card (lighter than base2 panel); tip ink base01.
-    tooltipPaper: "#fdf6e3",
-    tooltipInk: "#586e75",
-    tooltipBorder: "#93a1a1",
-  }),
-  // ggthemes theme_solarized_2 (light = FALSE): same geometry on the dark
-  // rebase — base02 panel (#073642), base03 grid (#002b36), base01 chrome.
-  // Accents are shared across light/dark by Solarized's design.
-  solarized_2dark: themed({
-    ink: "#586e75",
-    paper: "none",
-    panel: "#073642",
-    accent: "#268bd2",
-    grid: "#002b36",
-    axisText: "#586e75",
-    tickColor: "#586e75",
-    fontSize: 12,
-    axisTextSize: 12,
-    fontWeight: 400,
-    titleSize: 14.4,
-    subtitleSize: 12,
-    axisTitleSize: 12,
-    tickWidth: 0.5,
-    tickLength: 3.67,
-    gridWidth: 0.5,
-    ticksX: true,
-    ticksY: true,
-    // Tip package: base03 card (darker than base02 panel); tip ink base1.
-    tooltipPaper: "#002b36",
-    tooltipInk: "#93a1a1",
-    tooltipBorder: "#586e75",
-  }),
-  // ggthemes theme_wsj: Wall Street Journal chrome — "brown" paper (#f8f2e4),
-  // dotted black y-grid only, x axis line and ticks, no y line/ticks, big bold
-  // title (rel(2) of base 12). R's title_family = "mono" has no single-family
-  // token here, so the system sans stack carries both; axis.text bold and the
-  // blank axis.title also flatten into the shared roles. Accent is colors6 red
-  // so unmapped marks pair with the wsj palette.
-  wsj: themed({
-    ink: "#000000",
-    paper: "#f8f2e4",
-    panel: "#f8f2e4",
-    accent: "#c72e29",
-    grid: "#000000",
-    axisText: "#000000",
-    axisLine: "#000000",
-    tickColor: "#000000",
-    fontFamily: 'Arial, "Helvetica Neue", Helvetica, sans-serif',
-    fontSize: 12,
-    axisTextSize: 12,
-    fontWeight: 400,
-    titleSize: 24,
-    titleWeight: 700,
-    subtitleSize: 12,
-    subtitleWeight: 400,
-    axisTitleSize: 12,
-    axisTitleWeight: 400,
-    axisLineWidth: 0.5,
-    tickWidth: 0.5,
-    tickLength: 3.67,
-    gridWidth: 0.5,
-    gridDasharray: "1.5 3",
-    axisLineX: true,
-    ticksX: true,
-    gridX: false,
-  }),
-  // ggthemes theme_hc(style = "default"): Highcharts defaults — #D8D8D8
-  // y-only major grid, no panel border (panel background blank over white
-  // rect), theme_grey-lineage ticks and axis text. Accent is the first
-  // Highcharts default color.
-  hc: themed({
-    ink: "#000000",
-    paper: "#ffffff",
-    panel: "#ffffff",
-    accent: "#7cb5ec",
-    grid: "#D8D8D8",
-    axisText: "#4d4d4d",
-    tickColor: "#333333",
-    fontSize: 12,
-    axisTextSize: 9.6,
-    fontWeight: 400,
-    titleSize: 14.4,
-    subtitleSize: 12,
-    axisTitleSize: 12,
-    tickWidth: 0.5,
-    tickLength: 3.67,
-    gridWidth: 0.5,
-    ticksX: true,
-    ticksY: true,
-    gridX: false,
-  }),
-  // ggthemes theme_hc(style = "darkunica"): Highcharts dark theme — #2a2a2b
-  // paper, #707073 y-grid, #A0A0A3 text. R leaves axis.text at theme_grey's
-  // dark grey (dark-on-dark); this port uses the style's text colour for
-  // axis text and ticks (documented deviation). R's white title step
-  // flattens into the single ink role, as with the other ports.
-  hcdark: themed({
-    ink: "#A0A0A3",
-    paper: "#2a2a2b",
-    panel: "#2a2a2b",
-    accent: "#2b908f",
-    grid: "#707073",
-    axisText: "#A0A0A3",
-    tickColor: "#A0A0A3",
-    fontSize: 12,
-    axisTextSize: 9.6,
-    fontWeight: 400,
-    titleSize: 14.4,
-    subtitleSize: 12,
-    axisTitleSize: 12,
-    tickWidth: 0.5,
-    tickLength: 3.67,
-    gridWidth: 0.5,
-    ticksX: true,
-    ticksY: true,
-    gridX: false,
-    // Elevated neutral card; brighter tip ink; border stays grid-derived.
-    tooltipPaper: "#353538",
-    tooltipInk: "#E0E0E3",
-  }),
-  // ggthemes theme_pander() with its defaults (fc black, gc grey, gl dashed,
-  // boxes FALSE): white panel (pc "transparent" over white), dashed #bebebe
-  // grid both directions (R "grey" is #BEBEBE), grey ticks, bold 14.4 title,
-  // no visible panel border. Accent is the first pander palette color.
-  pander: themed({
-    ink: "#000000",
-    paper: "#ffffff",
-    panel: "#ffffff",
-    accent: "#56B4E9",
-    grid: "#bebebe",
-    axisText: "#000000",
-    tickColor: "#bebebe",
-    fontSize: 12,
-    axisTextSize: 9.6,
-    fontWeight: 400,
-    titleSize: 14.4,
-    titleWeight: 700,
-    subtitleSize: 12,
-    axisTitleSize: 12,
-    tickWidth: 0.25,
-    tickLength: 3.67,
-    gridWidth: 0.25,
-    gridDasharray: "3 3",
-    ticksX: true,
-    ticksY: true,
-  }),
-  // ggthemes theme_base: base-R graphics defaults — white panel with a black
-  // frame (panel.border; axis.line is blank in the theme_grey lineage), black
-  // ticks, no grid, black text, bold rel(1.2) title on base 16. Accent is
-  // black, matching base R's monochrome line/fill defaults.
-  base: themed({
-    ink: "#000000",
-    paper: "#ffffff",
-    panel: "#ffffff",
-    accent: "#000000",
-    grid: "none",
-    axisText: "#000000",
-    tickColor: "#000000",
-    panelBorder: "#000000",
-    fontSize: 16,
-    axisTextSize: 12.8,
-    fontWeight: 400,
-    titleSize: 19.2,
-    titleWeight: 700,
-    subtitleSize: 16,
-    axisTitleSize: 16,
-    tickWidth: 0.5,
-    tickLength: 8,
-    gridWidth: 0,
-    panelBorderWidth: 0.5,
-    ticksX: true,
-    ticksY: true,
-    gridX: false,
-    gridY: false,
-    showPanelBorder: true,
-  }),
-  // ggthemes theme_igray: the theme_gray inverse — white panel over a gray90
-  // (#e5e5e5) surround with a matching gray90 major grid, so the plot sits
-  // closer to the document's typographic color (Stata/Tableau style). Chrome
-  // follows this port's theme_grey lineage (#333/#4d4d4d).
-  igray: themed({
-    ink: "#333333",
-    paper: "#e5e5e5",
-    panel: "#ffffff",
-    grid: "#e5e5e5",
-    axisText: "#4d4d4d",
-    tickColor: "#333333",
-    fontSize: 12,
-    axisTextSize: 9.6,
-    fontWeight: 400,
-    titleSize: 14.4,
-    subtitleSize: 12,
-    axisTitleSize: 12,
-    tickWidth: 0.5,
-    tickLength: 3.67,
-    gridWidth: 0.5,
-    ticksX: true,
-    ticksY: true,
-  }),
-  // ggthemes theme_map: every axis/panel/grid element blank — marks only, for
-  // maps. Converges with `void` in this token model (both keep the title);
-  // base 9 typography documented from theme_bw(base_size = 9).
-  map: themed({
-    paper: "none",
-    panel: "none",
-    grid: "none",
-    fontSize: 9,
-    axisTextSize: 9,
-    axisTitleSize: 9,
-    gridWidth: 0,
-    gridX: false,
-    gridY: false,
-    axisLineX: false,
-    axisLineY: false,
-    ticksX: false,
-    ticksY: false,
-    labelsX: false,
-    labelsY: false,
-    showPanelBorder: false,
-    tickLength: 0,
-  }),
-  // ggthemes theme_solid(fill = NA): removes every non-geom element — no
-  // lines, no text, transparent background. This token model has no
-  // suppress-title role, so R's blanked title flattens into the shared
-  // void-like surface (documented); everything else matches.
-  solid: themed({
-    paper: "none",
-    panel: "none",
-    grid: "none",
-    fontSize: 12,
-    gridWidth: 0,
-    gridX: false,
-    gridY: false,
-    axisLineX: false,
-    axisLineY: false,
-    ticksX: false,
-    ticksY: false,
-    labelsX: false,
-    labelsY: false,
-    showPanelBorder: false,
-    tickLength: 0,
-  }),
-  // theme_test (#823): pinned high-contrast chrome for package tests / VR.
-  // Explicit literals — not an alias of light/classic — so product sweeps
-  // cannot retarget snapshot chrome. Stability over brand aesthetics.
-  test: themed({
-    ink: "#000000",
-    paper: "#ffffff",
-    panel: "#ffffff",
-    accent: "#000000",
-    grid: "#cccccc",
-    axisText: "#000000",
-    axisLine: "#000000",
-    tickColor: "#000000",
-    panelBorder: "#000000",
-    fontFamily: "Helvetica, Arial, sans-serif",
-    fontSize: 11,
-    axisTextSize: 12,
-    fontWeight: 400,
-    titleSize: 14,
-    titleWeight: 700,
-    subtitleSize: 12,
-    subtitleWeight: 400,
-    axisTitleSize: 11,
-    axisTitleWeight: 400,
-    captionSize: 9,
-    stripSize: 11,
-    stripWeight: 400,
-    axisLineWidth: 0.5,
-    tickWidth: 0.5,
-    tickLength: 4,
-    gridWidth: 0.5,
-    panelBorderWidth: 0.5,
-    gridDasharray: "",
-    axisLineX: true,
-    axisLineY: true,
-    ticksX: true,
-    ticksY: true,
-    gridX: true,
-    gridY: true,
-    showPanelBorder: true,
-  }),
+  ...CORE_BUILTIN_THEMES,
+  ...SOLARIZED_BUILTIN_THEMES,
+  ...PUBLICATION_BUILTIN_THEMES,
 });
 
-const LEGACY_BASE = themed(LEGACY_BASE_FOUNDATION);
-
-/** Edition-1 color themes with their original typography and chrome. */
-export const LEGACY_BUILTIN_THEMES: Readonly<Record<ThemeName, ThemeTokens>> = Object.freeze({
-  ...BUILTIN_THEMES,
-  default: LEGACY_BASE,
-  light: themed({
-    ...LEGACY_BASE_FOUNDATION,
-    ink: "#1f2328",
-    paper: "#ffffff",
-    panel: "none",
-    axisText: "#1f2328",
-    axisLine: "#1f2328",
-    tickColor: "#1f2328",
-    grid: "rgba(31,35,40,0.14)",
-  }),
-  dark: themed({
-    ...LEGACY_BASE_FOUNDATION,
-    ink: "#e6e8eb",
-    paper: "#16181d",
-    panel: "none",
-    accent: "#7ea1f0",
-    axisText: "#e6e8eb",
-    axisLine: "#e6e8eb",
-    tickColor: "#e6e8eb",
-    grid: "rgba(230,232,235,0.16)",
-    // Elevate fill only — border stays derived from legacy rgba grid.
-    tooltipPaper: "#22262d",
-  }),
-  minimal: themed({
-    ...LEGACY_BASE_FOUNDATION,
-    accent: "#9498a0",
-    grid: "rgba(128,128,128,0.12)",
-  }),
-});
+export const LEGACY_BUILTIN_THEMES: Readonly<Record<ThemeName, ThemeTokens>> =
+  legacyBuiltinThemes(BUILTIN_THEMES);
