+++ packages/core/tests/scales-m1/sequential-and-theme.test.ts
@@ -1,8 +1,6 @@
 /**
- * Sequential color ramps and theme registry (M1 scale surface companions).
+ * Sequential color ramps (M1 scale surface companions).
  */
-import { THEME_NAMES } from "@ggsvelte/spec";
-import { fromAny } from "@total-typescript/shoehorn";
 import { describe, expect, it } from "bun:test";
 
 import {
@@ -13,14 +11,6 @@ import {
   trainSequential,
   VIRIDIS_RAMP_10,
 } from "../../src/scales/color.ts";
-import { themed } from "../../src/theme-builtins.ts";
-import {
-  BUILTIN_THEMES,
-  LEGACY_BUILTIN_THEMES,
-  resolveTheme,
-  themeVar,
-  UnknownThemeError,
-} from "../../src/theme.ts";
 
 describe("sequential color", () => {
   it("interpolates the viridis ramp deterministically", () => {
@@ -86,395 +76,5 @@ describe("sequential color", () => {
     expect(logScale.colorOf(10)).toBe("#555555");
     expect(logScale.colorOf(100)).toBe("#aaaaaa");
     expect(logScale.colorOf(1000)).toBe("#ffffff");
-  });
-});
-
-describe("theme registry", () => {
-  const interactionColorRoles = [
-    "interactionInk",
-    "focusRing",
-    "crosshair",
-    "selectionFill",
-    "selectionStroke",
-    "tooltipPaper",
-    "tooltipInk",
-    "tooltipBorder",
-    "toolActive",
-  ] as const;
-
-  it("keeps every edition theme table complete against the spec registry", () => {
-    expect(Object.keys(BUILTIN_THEMES).toSorted()).toEqual([...THEME_NAMES].toSorted());
-    expect(Object.keys(LEGACY_BUILTIN_THEMES).toSorted()).toEqual([...THEME_NAMES].toSorted());
-  });
-
-  it("resolves the edition-2 typography and structural theme tokens", () => {
-    expect(resolveTheme()).toBe(BUILTIN_THEMES.default);
-    expect(resolveTheme("default").fontFamily).toContain("Roboto Condensed");
-    expect(resolveTheme("hrbr").axisLineX).toBe(false);
-    expect(resolveTheme("classic").axisLineX).toBe(true);
-    expect(resolveTheme("dark").paper).toBe("#16181d");
-    expect(resolveTheme("light").gridWidth).toBe(0.25);
-  });
-
-  it("resolves theme_grey / theme_gray as aliases of the ggplot2 grey-panel look (#824)", () => {
-    const ggplot2 = resolveTheme("ggplot2");
-    // Prefer shared token maps (not a re-skin): same object identity as ggplot2.
-    expect(resolveTheme("grey")).toBe(ggplot2);
-    expect(resolveTheme("gray")).toBe(ggplot2);
-    expect(BUILTIN_THEMES.grey).toBe(BUILTIN_THEMES.ggplot2);
-    expect(BUILTIN_THEMES.gray).toBe(BUILTIN_THEMES.ggplot2);
-    // Grey-panel signature (not the hrbr/default white panel).
-    expect(ggplot2.panel).toBe("#ebebeb");
-    expect(ggplot2.grid.toLowerCase()).toBe("#ffffff");
-  });
-
-  it("registers theme_test as a pinned high-contrast snapshot theme (#823)", () => {
-    // Stability contract: fixed literals so product-theme sweeps (#753-style)
-    // cannot silently retarget the test chrome. Not an alias of light/classic.
-    expect(THEME_NAMES).toContain("test");
-    expect(THEME_NAMES.at(-1)).toBe("test");
-    const tokens = resolveTheme("test");
-    expect(tokens.paper).toBe("#ffffff");
-    expect(tokens.panel).toBe("#ffffff");
-    expect(tokens.ink).toBe("#000000");
-    expect(tokens.accent).toBe("#000000");
-    expect(tokens.grid).toBe("#cccccc");
-    expect(tokens.axisText).toBe("#000000");
-    expect(tokens.axisLine).toBe("#000000");
-    expect(tokens.tickColor).toBe("#000000");
-    expect(tokens.panelBorder).toBe("#000000");
-    expect(tokens.fontFamily).toBe("Helvetica, Arial, sans-serif");
-    expect(tokens.fontSize).toBe(11);
-    expect(tokens.axisTextSize).toBe(12);
-    expect(tokens.titleSize).toBe(14);
-    expect(tokens.subtitleSize).toBe(12);
-    expect(tokens.axisTitleSize).toBe(11);
-    expect(tokens.captionSize).toBe(9);
-    expect(tokens.stripSize).toBe(11);
-    expect(tokens.axisLineWidth).toBe(0.5);
-    expect(tokens.tickWidth).toBe(0.5);
-    expect(tokens.tickLength).toBe(4);
-    expect(tokens.gridWidth).toBe(0.5);
-    expect(tokens.panelBorderWidth).toBe(0.5);
-    expect(tokens.ticksX).toBe(true);
-    expect(tokens.ticksY).toBe(true);
-    expect(tokens.gridX).toBe(true);
-    expect(tokens.gridY).toBe(true);
-    expect(tokens.axisLineX).toBe(true);
-    expect(tokens.axisLineY).toBe(true);
-    expect(tokens.showPanelBorder).toBe(true);
-    // Not a silent alias of product themes that still evolve.
-    expect(tokens.fontFamily).not.toBe(resolveTheme("default").fontFamily);
-    expect(tokens.grid).not.toBe(resolveTheme("classic").grid);
-  });
-
-  it("registers theme_bw as a white-panel print theme distinct from light and ggplot2 (#820)", () => {
-    // Independent contract from ggplot2's published complete-theme role:
-    // white panel + grey grid + rectangular border (not grey-panel/white-grid
-    // ggplot2, and not light's thinner/lighter chrome).
-    expect(THEME_NAMES).toContain("bw");
-    const bw = resolveTheme("bw");
-    const light = resolveTheme("light");
-    const ggplot2 = resolveTheme("ggplot2");
-
-    expect(bw.paper).toBe("#ffffff");
-    expect(bw.panel).toBe("#ffffff");
-    expect(bw.grid).not.toBe("none");
-    expect(bw.grid).not.toBe("#ffffff");
-    expect(bw.showPanelBorder).toBe(true);
-    expect(bw.ticksX).toBe(true);
-    expect(bw.ticksY).toBe(true);
-    expect(bw.axisLineX).toBe(false);
-    expect(bw.axisLineY).toBe(false);
-    expect(bw.axisTextSize).toBeGreaterThanOrEqual(12);
-    expect(bw.panelBorder).toBe("#333333");
-    expect(bw.grid).toBe("#e5e5e5");
-
-    // Distinct from cousins that already ship.
-    expect(bw.panel).not.toBe(ggplot2.panel);
-    expect(bw.grid).not.toBe(ggplot2.grid);
-    expect(bw.panelBorder).not.toBe(light.panelBorder);
-    expect(bw.gridWidth).not.toBe(light.gridWidth);
-  });
-
-  it("resolves theme_linedraw as high-contrast black-line chrome (#821)", () => {
-    const t = resolveTheme("linedraw");
-    const classic = resolveTheme("classic");
-    const ggplot2 = resolveTheme("ggplot2");
-
-    // White panel (not ggplot2 grey); black monochrome chrome.
-    expect(t.panel).toBe("#ffffff");
-    expect(t.paper).toBe("#ffffff");
-    expect(t.panel).not.toBe(ggplot2.panel);
-    expect(t.ink.toLowerCase()).toBe("#000000");
-    expect(t.grid.toLowerCase()).toBe("#000000");
-    expect(t.panelBorder.toLowerCase()).toBe("#000000");
-    expect(t.tickColor.toLowerCase()).toBe("#000000");
-    expect(t.axisText.toLowerCase()).toBe("#000000");
-    // Grid present (classic suppresses grid entirely).
-    expect(t.gridX).toBe(true);
-    expect(t.gridY).toBe(true);
-    expect(classic.gridX).toBe(false);
-    expect(t.showPanelBorder).toBe(true);
-    expect(t.ticksX).toBe(true);
-    expect(t.ticksY).toBe(true);
-    expect(t.axisTextSize).toBeGreaterThanOrEqual(12);
-    // Hairline black grid — not solid graph paper.
-    expect(t.gridWidth).toBeLessThanOrEqual(0.35);
-  });
-
-  it("keeps axis tick labels readable on light/minimal family themes (#753)", () => {
-    // 8.8px was unreadable next to 15px titles and ~12–16px tooltips on the
-    // docs homepage hero. Floor is intentionally above 11 so axis chrome is
-    // not fine print at 640×400.
-    for (const name of [
-      "light",
-      "minimal",
-      "ggplot2",
-      "classic",
-      "few",
-      "bw",
-      "linedraw",
-      "grey",
-      "gray",
-      "test",
-    ] as const) {
-      expect(resolveTheme(name).axisTextSize, name).toBeGreaterThanOrEqual(12);
-    }
-    expect(resolveTheme("default").axisTitleSize).toBeGreaterThanOrEqual(11);
-  });
-
-  it("object themes override roles over a named base", () => {
-    const tokens = resolveTheme({
-      name: "dark",
-      accent: "#ff0000",
-      focusRing: "#00ff00",
-      interactionMuted: 0.5,
-      tooltipPaper: "#111111",
-    });
-    expect(tokens.accent).toBe("#ff0000");
-    expect(tokens.ink).toBe(BUILTIN_THEMES.dark.ink);
-    expect(tokens.focusRing).toBe("#00ff00");
-    expect(tokens.interactionMuted).toBe(0.5);
-    expect(tokens.tooltipPaper).toBe("#111111");
-  });
-
-  it("unknown names throw (tier-1 error) and themeVar wraps --gg-* vars", () => {
-    expect(() => resolveTheme(fromAny("darkk"))).toThrow(UnknownThemeError);
-    expect(themeVar("ink", BUILTIN_THEMES.default)).toBe("var(--gg-ink, #262626)");
-  });
-
-  it("resolves the complete interaction visual language as CSS-variable roles", () => {
-    for (const name of Object.keys(BUILTIN_THEMES) as (keyof typeof BUILTIN_THEMES)[]) {
-      const tokens = resolveTheme(name);
-      for (const role of interactionColorRoles) {
-        expect(tokens[role], `${name}.${role}`).toBeTruthy();
-        expect(themeVar(role, tokens), `${name}.${role} CSS variable`).toBe(
-          `var(--gg-${role}, ${tokens[role]})`,
-        );
-      }
-      expect(tokens.interactionMuted, `${name}.interactionMuted`).toBeGreaterThan(0);
-      expect(tokens.interactionMuted, `${name}.interactionMuted`).toBeLessThan(1);
-      expect(themeVar("interactionMuted", tokens)).toBe(
-        `var(--gg-interactionMuted, ${tokens.interactionMuted})`,
-      );
-    }
-
-    expect(BUILTIN_THEMES.default.tooltipPaper).toBe(BUILTIN_THEMES.default.paper);
-    expect(BUILTIN_THEMES.default.tooltipInk).toBe(BUILTIN_THEMES.default.ink);
-    // dark elevates tip paper above chart surface (see elevated problem-theme tests).
-    expect(BUILTIN_THEMES.dark.tooltipInk).toBe(BUILTIN_THEMES.dark.ink);
-    expect(BUILTIN_THEMES.default.selectionFill).toContain("rgba(");
-    expect(BUILTIN_THEMES.dark.selectionFill).toContain("rgba(");
-  });
-
-  // #1069 — flat tooltip chrome only for unframed gridless themes (tufte/void).
-  // Framed gridless themes (few/classic) keep panelBorder keylines.
-  it("unframed gridless themes derive a flat tooltip keyline; framed ones keep panelBorder", () => {
-    expect(BUILTIN_THEMES.tufte.grid).toBe("none");
-    expect(BUILTIN_THEMES.tufte.tooltipBorder).toBe("transparent");
-    expect(BUILTIN_THEMES.tufte.tooltipBorder).not.toBe(BUILTIN_THEMES.tufte.panelBorder);
-    expect(BUILTIN_THEMES.void.grid).toBe("none");
-    expect(BUILTIN_THEMES.void.tooltipBorder).toBe("transparent");
-
-    // few: gridless but showPanelBorder — keep the panel border keyline.
-    expect(BUILTIN_THEMES.few.grid).toBe("none");
-    expect(BUILTIN_THEMES.few.showPanelBorder).toBe(true);
-    expect(BUILTIN_THEMES.few.tooltipBorder).toBe(BUILTIN_THEMES.few.panelBorder);
-    expect(BUILTIN_THEMES.few.tooltipBorder).not.toBe("transparent");
-
-    // classic: gridless but axis lines — keep the panel border keyline.
-    expect(BUILTIN_THEMES.classic.grid).toBe("none");
-    expect(BUILTIN_THEMES.classic.axisLineX).toBe(true);
-    expect(BUILTIN_THEMES.classic.tooltipBorder).toBe(BUILTIN_THEMES.classic.panelBorder);
-    expect(BUILTIN_THEMES.classic.tooltipBorder).not.toBe("transparent");
-
-    // Themes that still draw a grid keep a visible hairline from the grid color.
-    expect(BUILTIN_THEMES.default.tooltipBorder).toBe(BUILTIN_THEMES.default.grid);
-    expect(BUILTIN_THEMES.default.tooltipBorder).not.toBe("transparent");
-    expect(BUILTIN_THEMES.light.tooltipBorder).toBe(BUILTIN_THEMES.light.grid);
-
-    // Object overrides still win over the derivation.
-    expect(resolveTheme({ name: "tufte", tooltipBorder: "#aabbcc" }).tooltipBorder).toBe("#aabbcc");
-  });
-
-  it("themed() accepts tooltip role overrides", () => {
-    const tokens = themed({
-      paper: "#111111",
-      ink: "#eeeeee",
-      tooltipPaper: "#222222",
-      tooltipInk: "#ffffff",
-      tooltipBorder: "#444444",
-    });
-    expect(tokens.tooltipPaper).toBe("#222222");
-    expect(tokens.tooltipInk).toBe("#ffffff");
-    expect(tokens.tooltipBorder).toBe("#444444");
-  });
-
-  it("object path sticky-when-elevated: pure themes still re-derive tip from paper", () => {
-    const tokens = resolveTheme({ name: "default", paper: "#ffeeee" });
-    expect(tokens.tooltipPaper).toBe("#ffeeee");
-  });
-
-  it("object path sticky-when-elevated: elevated base tips stick without explicit tip*", () => {
-    // Synthetic complete theme: pure derivation would set tip = paper, but the
-    // frozen entry elevates the tip package above that derivation.
-    const elevated = themed({
-      paper: "#16181d",
-      panel: "#16181d",
-      ink: "#e6e8eb",
-      grid: "#3b3f46",
-      tooltipPaper: "#22262d",
-      tooltipInk: "#ffffff",
-      tooltipBorder: "#555555",
-    });
-    const builtins = { ...BUILTIN_THEMES, default: elevated };
-    const viaString = resolveTheme("default", builtins);
-    const viaObject = resolveTheme({ name: "default", titleSize: 20 }, builtins);
-    const viaAccentOnly = resolveTheme({ name: "default", accent: "#ff0000" }, builtins);
-
-    expect(viaString.tooltipPaper).toBe("#22262d");
-    expect(viaObject.tooltipPaper).toBe("#22262d");
-    expect(viaObject.tooltipInk).toBe("#ffffff");
-    expect(viaObject.tooltipBorder).toBe("#555555");
-    expect(viaAccentOnly.tooltipPaper).toBe("#22262d");
-    // Explicit tip still wins over sticky elevation.
-    expect(resolveTheme({ name: "default", tooltipPaper: "#111111" }, builtins).tooltipPaper).toBe(
-      "#111111",
-    );
-  });
-
-  it("object path sticky-when-elevated: few/classic keep panelBorder keyline (no false-sticky)", () => {
-    // Guards incomplete pureFromBase foundation lists that would re-derive
-    // transparent borders for framed gridless themes.
-    expect(resolveTheme({ name: "few", fontSize: 14 }).tooltipBorder).toBe(
-      BUILTIN_THEMES.few.panelBorder,
-    );
-    expect(resolveTheme({ name: "classic", fontSize: 14 }).tooltipBorder).toBe(
-      BUILTIN_THEMES.classic.panelBorder,
-    );
-    expect(resolveTheme({ name: "tufte", fontSize: 14 }).tooltipBorder).toBe("transparent");
-  });
-
-  it("LEGACY light/dark/minimal tip contracts (no stale tip leak from foundation spread)", () => {
-    // LEGACY dark elevates paper only; border tracks legacy rgba grid.
-    expect(LEGACY_BUILTIN_THEMES.dark.tooltipPaper).toBe("#22262d");
-    expect(LEGACY_BUILTIN_THEMES.dark.tooltipPaper).not.toBe(LEGACY_BUILTIN_THEMES.dark.paper);
-    expect(LEGACY_BUILTIN_THEMES.dark.tooltipInk).toBe(LEGACY_BUILTIN_THEMES.dark.ink);
-    expect(LEGACY_BUILTIN_THEMES.dark.tooltipBorder).toBe(LEGACY_BUILTIN_THEMES.dark.grid);
-    expect(LEGACY_BUILTIN_THEMES.dark.tooltipBorder).toBe("rgba(230,232,235,0.16)");
-
-    expect(LEGACY_BUILTIN_THEMES.light.tooltipPaper).toBe(LEGACY_BUILTIN_THEMES.light.paper);
-    expect(LEGACY_BUILTIN_THEMES.light.tooltipBorder).toBe(LEGACY_BUILTIN_THEMES.light.grid);
-
-    expect(LEGACY_BUILTIN_THEMES.minimal.tooltipBorder).toBe(LEGACY_BUILTIN_THEMES.minimal.grid);
-    expect(LEGACY_BUILTIN_THEMES.minimal.tooltipBorder).toBe("rgba(128,128,128,0.12)");
-  });
-
-  it("LEGACY light/dark keep transparent letterbox gutters (edition-1 paper overrides)", () => {
-    // paper is opaque, but letterboxFill stays "none" so fixed-aspect margins
-    // remain see-through over the host page (prior LEGACY_BASE behavior).
-    expect(LEGACY_BUILTIN_THEMES.light.paper).toBe("#ffffff");
-    expect(LEGACY_BUILTIN_THEMES.light.letterboxFill).toBe("none");
-    expect(LEGACY_BUILTIN_THEMES.dark.paper).toBe("#16181d");
-    expect(LEGACY_BUILTIN_THEMES.dark.letterboxFill).toBe("none");
-    expect(LEGACY_BUILTIN_THEMES.minimal.letterboxFill).toBe("none");
-  });
-
-  it("Solarized family elevates tooltip cards above the reading surface", () => {
-    const PROBLEM_SOLARIZED = [
-      "solarized",
-      "solarizeddark",
-      "solarized_2",
-      "solarized_2dark",
-    ] as const;
-
-    for (const name of PROBLEM_SOLARIZED) {
-      const tokens = BUILTIN_THEMES[name];
-      const readingSurface = tokens.paper === "none" ? tokens.panel : tokens.paper;
-      const viaString = resolveTheme(name);
-      const viaObject = resolveTheme({ name, titleSize: 20 });
-
-      expect(viaString.tooltipPaper, name).toBe(tokens.tooltipPaper);
-      expect(viaObject.tooltipPaper, name).toBe(tokens.tooltipPaper);
-      expect(tokens.tooltipPaper, name).not.toBe(readingSurface);
-      // LEGACY spreads edition-2 Solarized tips.
-      expect(LEGACY_BUILTIN_THEMES[name].tooltipPaper, name).toBe(tokens.tooltipPaper);
-    }
-
-    // Exact Schoonover anchors (tip ink stronger than axis ink on low-contrast bases).
-    expect(BUILTIN_THEMES.solarized.tooltipPaper).toBe("#eee8d5"); // base2
-    expect(BUILTIN_THEMES.solarized.tooltipInk).toBe("#586e75"); // base01
-    expect(BUILTIN_THEMES.solarized.tooltipBorder).toBe("#93a1a1"); // base1
-    expect(BUILTIN_THEMES.solarized.tooltipInk).not.toBe(BUILTIN_THEMES.solarized.ink);
-
-    expect(BUILTIN_THEMES.solarized_2.tooltipPaper).toBe("#fdf6e3"); // base3
-    expect(BUILTIN_THEMES.solarized_2.tooltipInk).toBe("#586e75");
-    expect(BUILTIN_THEMES.solarized_2.tooltipBorder).toBe("#93a1a1");
-
-    expect(BUILTIN_THEMES.solarizeddark.tooltipPaper).toBe("#073642"); // base02
-    expect(BUILTIN_THEMES.solarizeddark.tooltipInk).toBe("#93a1a1"); // base1
-    expect(BUILTIN_THEMES.solarizeddark.tooltipBorder).toBe("#586e75"); // base01
-    expect(BUILTIN_THEMES.solarizeddark.tooltipInk).not.toBe(BUILTIN_THEMES.solarizeddark.ink);
-
-    expect(BUILTIN_THEMES.solarized_2dark.tooltipPaper).toBe("#002b36"); // base03
-    expect(BUILTIN_THEMES.solarized_2dark.tooltipInk).toBe("#93a1a1");
-    expect(BUILTIN_THEMES.solarized_2dark.tooltipBorder).toBe("#586e75");
-  });
-
-  it("elevates tooltip cards on dark, hcdark, fivethirtyeight, and economist", () => {
-    // Edition-2 dark: paper only; ink/border stay derived.
-    expect(BUILTIN_THEMES.dark.tooltipPaper).toBe("#22262d");
-    expect(BUILTIN_THEMES.dark.tooltipPaper).not.toBe(BUILTIN_THEMES.dark.paper);
-    expect(BUILTIN_THEMES.dark.tooltipInk).toBe(BUILTIN_THEMES.dark.ink);
-    expect(BUILTIN_THEMES.dark.tooltipBorder).toBe(BUILTIN_THEMES.dark.grid);
-    expect(resolveTheme({ name: "dark", accent: "#ff0000" }).tooltipPaper).toBe("#22262d");
-
-    // hcdark: elevated fill + brighter tip ink; border tracks grid.
-    expect(BUILTIN_THEMES.hcdark.tooltipPaper).toBe("#353538");
-    expect(BUILTIN_THEMES.hcdark.tooltipInk).toBe("#E0E0E3");
-    expect(BUILTIN_THEMES.hcdark.tooltipBorder).toBe(BUILTIN_THEMES.hcdark.grid);
-    expect(resolveTheme({ name: "hcdark", titleSize: 20 }).tooltipPaper).toBe("#353538");
-
-    // fivethirtyeight: white tip + soft grey border.
-    expect(BUILTIN_THEMES.fivethirtyeight.tooltipPaper).toBe("#ffffff");
-    expect(BUILTIN_THEMES.fivethirtyeight.tooltipInk).toBe(BUILTIN_THEMES.fivethirtyeight.ink);
-    expect(BUILTIN_THEMES.fivethirtyeight.tooltipBorder).toBe("#d0d0d0");
-    expect(resolveTheme({ name: "fivethirtyeight", titleSize: 20 }).tooltipPaper).toBe("#ffffff");
-
-    // economist: same-hue lighter tip + tick-color border.
-    expect(BUILTIN_THEMES.economist.tooltipPaper).toBe("#eef5f8");
-    expect(BUILTIN_THEMES.economist.tooltipInk).toBe(BUILTIN_THEMES.economist.ink);
-    expect(BUILTIN_THEMES.economist.tooltipBorder).toBe("#6794a7");
-    expect(resolveTheme({ name: "economist", titleSize: 20 }).tooltipPaper).toBe("#eef5f8");
-
-    // LEGACY non-dark problem names equal edition-2 tips via spread.
-    for (const name of ["hcdark", "fivethirtyeight", "economist"] as const) {
-      expect(LEGACY_BUILTIN_THEMES[name].tooltipPaper).toBe(BUILTIN_THEMES[name].tooltipPaper);
-    }
-
-    // Pure controls still hold.
-    expect(BUILTIN_THEMES.default.tooltipPaper).toBe(BUILTIN_THEMES.default.paper);
-    expect(BUILTIN_THEMES.ggplot2.tooltipPaper).toBe(BUILTIN_THEMES.ggplot2.paper);
   });
 });
